first PR review, moved SimulatorBase to own file, backend imports only on target call
