+++ tangelo/helpers/utils.py
@@ -19,6 +19,7 @@ a default simulator.
 
 import os
 import sys
+from importlib import util
 
 
 class HiddenPrints:
@@ -34,13 +35,9 @@ class HiddenPrints:
 
 
 def is_package_installed(package_name):
-    try:
-        exec(f"import {package_name}")
-        # DEBUG print(f'{package_name:16s} :: found')
-        return True
-    except ModuleNotFoundError:
-        # DEBUG print(f'{package_name:16s} :: not found')
-        return False
+    """Check if module is installed without importing."""
+    spam_spec = util.find_spec(package_name)
+    return spam_spec is not None
 
 
 # List all backends and statevector ones supported by Simulator class
